Convert a single object into a list of target type
simple test passing with as per the basic requirement of gh-3.

We just need tidy up the delegate enrollment and move it to a seperate
class. We also need to perform a few tests on the abstraction.

Furthermore, we will need to ensure that we are not enrolling against an
output type of IEnumerable.